Refact
Separate components confir, confirm Datas.
